+++ backend/src/main/java/com/andreea/ticket_tracker/services/ProjectService.java
@@ -11,7 +11,9 @@ import com.andreea.ticket_tracker.mapper.UserDTOMapper;
 import com.andreea.ticket_tracker.mapper.ProjectDTOMapper;
 import com.andreea.ticket_tracker.repository.ProjectRepository;
 import com.andreea.ticket_tracker.repository.UserRepository;
+import com.andreea.ticket_tracker.security.config.ProjectSecurityEvaluator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,12 +24,14 @@ public class ProjectService {
     private final ProjectRepository projectRepository;
     private final UserRepository userRepository;
     private final UserDTOMapper userDTOMapper;
+    private final ProjectSecurityEvaluator projectSecurity;
 
     @Autowired
-    public ProjectService(ProjectRepository projectRepository, UserRepository userRepository, UserDTOMapper userDTOMapper) {
+    public ProjectService(ProjectRepository projectRepository, UserRepository userRepository, UserDTOMapper userDTOMapper, ProjectSecurityEvaluator projectSecurity) {
         this.projectRepository = projectRepository;
         this.userRepository = userRepository;
         this.userDTOMapper = userDTOMapper;
+        this.projectSecurity = projectSecurity;
     }
 
     public ProjectResponseDTO createProject(ProjectRequestDTO dto){
@@ -37,8 +41,13 @@ public class ProjectService {
     }
 
     public List<ProjectResponseDTO> getAllProjects(){
-        return projectRepository.findAll()
-                .stream()
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        List<Project> projects = projectSecurity.isUserAdmin()
+                ? projectRepository.findAll()
+                : projectRepository.findAllByUsers_Username(currentUsername);
+
+        return projects.stream()
                 .map(ProjectDTOMapper::toDTO)
                 .toList();
     }
@@ -47,6 +56,8 @@ public class ProjectService {
         Project project = projectRepository.findById(id)
                 .orElseThrow(ProjectNotFoundException::new);
 
+        projectSecurity.validateUserAccess(project);
+
         return ProjectDTOMapper.toDTO(project);
     }
 
@@ -54,6 +65,7 @@ public class ProjectService {
         Project project = projectRepository.findById(id)
                 .orElseThrow(ProjectNotFoundException::new);
 
+        projectSecurity.validateUserAccess(project);
         project.setName(dto.getName());
         project.setDescription(dto.getDescription());
 
@@ -62,15 +74,18 @@ public class ProjectService {
     }
 
     public void deleteProject(Long id){
-        projectRepository.findById(id)
+        Project project = projectRepository.findById(id)
                 .orElseThrow(ProjectNotFoundException::new);
 
+        projectSecurity.validateUserAccess(project);
         projectRepository.deleteById(id);
     }
 
     public void assignUserToProject(Long projectId, Long userId){
         Project project = projectRepository.findById(projectId)
                 .orElseThrow(ProjectNotFoundException::new);
+
+        projectSecurity.validateUserAccess(project);
 
         User user = userRepository.findById(userId)
                 .orElseThrow(UserNotFoundException::new);
@@ -80,8 +95,10 @@ public class ProjectService {
     }
 
     public List<UserResponseDTO> getProjectMembers(Long projectId){
-        projectRepository.findById(projectId)
+        Project project = projectRepository.findById(projectId)
                 .orElseThrow(ProjectNotFoundException::new);
+
+        projectSecurity.validateUserAccess(project);
 
         return userRepository.findAllByProjects_Id(projectId)
                 .stream()
@@ -92,6 +109,9 @@ public class ProjectService {
     public void removeUserFromProject(Long projectId, Long userId) {
         Project project = projectRepository.findById(projectId)
                 .orElseThrow(ProjectNotFoundException::new);
+
+        projectSecurity.validateUserAccess(project);
+
         User user = userRepository.findById(userId)
                 .orElseThrow(UserNotFoundException::new);
